+++ biengual-core/src/main/java/com/biengual/core/annotation/WebCrawler.java
@@ -1,0 +1,20 @@
+package com.biengual.core.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import org.springframework.stereotype.Component;
+
+/**
+ * Web Crawler 임을 명시하기 위한 어노테이션
+ *
+ * @author 김영래
+ */
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@Component
+public @interface WebCrawler {
+    String value() default "";
+}
