+++ user-api/src/main/java/com/biengual/userapi/crawling/infrastructure/CrawlingReaderImpl.java
@@ -26,6 +26,7 @@ public class CrawlingReaderImpl implements CrawlingReader {
     public List<ContentCommand.CrawlingContent> getDailyUrlsForCrawling() {
         List<ContentCommand.CrawlingContent> commands = new ArrayList<>();
 
+        // TODO: URI 리스트 ENUM 등의 클래스로 분리
         // 1-1. CNN 페이지 의 MORE TOP STORIES
         Document doc = null;
         try {
@@ -36,7 +37,7 @@ public class CrawlingReaderImpl implements CrawlingReader {
 
         Elements elements =
             doc.select("[data-collapsed-text='More top stories'] .container_lead-plus-headlines__cards-wrapper");
-        for (Element element : elements) {
+        for (Element element : elements) {  // 5~6개 (영상 포함된 기사가 1개 정도 있어서 크롤링 하지 않는 기사 있음)
             commands.add(
                 ContentCommand.CrawlingContent.builder()
                     .url(Objects.requireNonNull(element.select("a").first()).attr("href"))
@@ -46,6 +47,14 @@ public class CrawlingReaderImpl implements CrawlingReader {
         }
 
         // 1-2. Youtube 채널의 rss
+        // 채널 별로 매일 2개 씩 크롤링
+        String youtube1 = URI.create("https://www.youtube.com/@TEDEd").toString();
+        String youtube2 = URI.create("https://www.youtube.com/@Vox").toString();
+        String youtube3 = URI.create("https://www.youtube.com/@lifenoggin").toString();
+
+
+
+
 
         return commands;
     }
